+++ packages/nuxt-module/package.json
@@ -15,6 +15,9 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@nuxt/types": "^0.6.4",
+    "@nuxt/typescript-build": "^0.6.0",
+    "@nuxt/typescript-runtime": "^0.4.0",
     "@nuxt/utils": "^2.11.0",
     "cookie-universal": "^2.1.1",
     "fs-jetpack": "^2.2.3"
