+++ packages/nuxt-module/src/@nuxt/utils.d.ts
@@ -1,0 +1,6 @@
+// just POC; move into right place afterwards.
+
+declare module '@nuxt/utils' {
+  export function sortRoutes(any:any): any;
+  export function createRoutes(any:any): any;
+}
